# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Eventless API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/events": {
      "get": {
        "summary": "Retrieves events where registration is possible",
        "description": "This endpoint returns a list of events where registration is possible. It requires a valid API key for authorization.",
        "parameters": [
          {
            "in": "header",
            "name": "api-key",
            "required": true,
            "description": "API key for authorization",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized, invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "UNAUTHORIZED"
                    },
                    "message": {
                      "type": "string",
                      "example": "Your API Key is invalid."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "INTERNAL_SERVER_ERROR"
                    },
                    "message": {
                      "type": "string",
                      "example": "please wait a moment and try again"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}